Revert "Silence conversion warnings"
This reverts commit 17dcb730a96e60b206e8fe7461964c6745a07f97.

+++ include/box2d/b2_growable_stack.h
@@ -56,8 +56,8 @@ public:
 		{
 			T* old = m_stack;
 			m_capacity *= 2;
-			m_stack = (T*)b2Alloc(m_capacity * static_cast<int32>(sizeof(T)));
-			memcpy(m_stack, old, static_cast<size_t>(m_count) * sizeof(T));
+			m_stack = (T*)b2Alloc(m_capacity * sizeof(T));
+			memcpy(m_stack, old, m_count * sizeof(T));
 			if (old != m_array)
 			{
 				b2Free(old);
